+++ packages/data-adapters/src/index.ts
@@ -11,6 +11,7 @@ export * from './cache';
 export * from './seed';
 export * from './Provider';
 export * from './MockProvider';
+export * from './BinanceProvider';
 export * from './providerRegistry';
 export * from './conformance';
 export { YahooProvider } from './stubs/YahooProvider';